+++ packages/examples/src/common/node/server-commons.ts
@@ -4,10 +4,9 @@
  * ------------------------------------------------------------------------------------------ */
 import { WebSocketServer, type ServerOptions } from 'ws';
 import { IncomingMessage, Server } from 'node:http';
-import { URL } from 'node:url';
 import { Socket } from 'node:net';
 import { dirname } from 'node:path';
-import { fileURLToPath } from 'node:url';
+import { fileURLToPath, URL } from 'node:url';
 import * as cp from 'node:child_process';
 import { type IWebSocket, WebSocketMessageReader, WebSocketMessageWriter } from 'vscode-ws-jsonrpc';
 import { createConnection, createServerProcess, forward } from 'vscode-ws-jsonrpc/server';