+++ packages/wrapper-react/CHANGELOG.md
@@ -4,8 +4,9 @@ All notable changes to npm module [@typefox/monaco-editor-react](https://www.npm
 
 ## [7.8.0] - unreleased
 
-- Updated all `@codingame/monaco-vscode` packages to `26.0.1`.
+- Updated all `@codingame/monaco-vscode` packages to `26.2.1`.
 - Updated to `monaco-languageclient@10.8.0`.
+- Dropped eslint and rely fully on oxlint.
 
 ## [7.7.0] - 2026-02-04